+++ lib/lang/src/compile/mod.rs
@@ -1,5 +1,5 @@
 mod error;
-mod expressions;
+mod expression_compilation;
 mod interfaces;
 mod list_type_configuration;
 mod module_compilation;
